// web-admin/src/generated/api-schema.d.ts
export interface components {
    schemas: {
// ...
        LiveRequest: {
            /** Format: int64 */
            channelId: number | string;
            /**
             * Format: int32
             * @default 2
             */
            streamType: number | string;
            /** @default browser */
            profile: string;
        };
        LoginRequest: {
            username: string;
            password: string;
            /** @default web */
            clientType: string;
            /** @default 2.0.0 */
            clientVersion: string;
        };
// ...
        PlaybackRequest: {
            /** Format: int64 */
            channelId: number | string;
            /** Format: date-time */
            start: string;
            /** Format: date-time */
            end: string;
            /** @default browser */
            profile: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}